revert commit e9bcede ... unmerge navbar-toolbar update styles & test
