+++ src/components/navbar/style.styl
@@ -1,0 +1,34 @@
+.navbar
+    background-color: var(--Navbar-bg)
+    box-shadow: 0 3px 5px -5px rgba(0, 0, 0, 15)
+    align-items: center
+    display: flex
+    justify-content: space-between
+    height: var(--Navbar-height)
+    overflow: hidden
+    user-select: none
+    width: 100%
+    z-index: 10
+
+
+    .menu
+        display: flex
+        align-items: center
+        justify-content: center
+        margin: 0 15px
+
+        .title, .button
+            font-size: 20px
+            margin-right: 15px
+
+        .title
+            font-family: var(--Font-title)
+            font-weight: bold
+            text-transform: uppercase
+        .button
+            opacity: 0.75
+            cursor: pointer
+            transition: opacity 0.25s
+
+            &:hover
+                opacity: 1
