+++ src/components/toolbar/style.styl
@@ -1,76 +1,41 @@
 .toolbar
-    background-color: var(--Navbar-bg)
-    box-shadow: 0 3px 5px -5px rgba(0, 0, 0, 15)
     align-items: center
+    color: #FFF
     display: flex
-    justify-content: space-between
-    height: var(--Navbar-height)
-    overflow: hidden
-    user-select: none
-    width: 100%
-    z-index: 10
+    margin: 0 15px
+    
+    .button
+        border-radius: 3px
+        cursor: pointer
+        display: inline-block
+        margin-left: 15px
+        opacity: 0.65
+        transition: opacity 0.25s
 
+        &.active
+            background-color: rgba(225,225,225,0.15);
 
-    .menu
-        display: flex
-        align-items: center
-        justify-content: center
-        margin-left: 10px
+        &.active,
+        &:hover
+            opacity: 1
 
-        .title, .button
-            font-size: 20px
-            margin-right: 15px
+        &.update
+            i:before
+                background-color: var(--Main-color)
+                border-radius: 5px
+                content: ''
+                display: block
+                height: 8px
+                position: absolute
+                right: -2px
+                top: -2px
+                width: 8px
+                z-index: 99
 
-        .title
-            font-family: var(--Font-title)
-            font-weight: bold
-            text-transform: uppercase
-        .button
-            opacity: 0.75
-            cursor: pointer
-            transition: opacity 0.25s
-
-            &:hover
-                opacity: 1
-
-    .actions
-        align-items: center
-        color: #FFF
-        display: flex
-        margin-right: 10px
-
-        .button
-            border-radius: 3px
-            cursor: pointer
-            display: inline-block
-            margin-left: 15px
-            opacity: 0.65
-            transition: opacity 0.25s
-
-            &.active
-                background-color: rgba(225,225,225,0.15);
-
-            &.active
-            &:hover
-                opacity: 1
-
-            &.update
-                i:before
-                    background-color: var(--Main-color)
-                    border-radius: 5px
-                    content: ''
-                    display: block
-                    height: 8px
-                    position: absolute
-                    right: -2px
-                    top: -2px
-                    width: 8px
-                    z-index: 99
-
-            i
-                font-size 22px
-                height: 22px
-                line-height: 22px
-                padding: 5px;
-                position: relative
-                width: 22px
+        i
+            font-size 22px
+            height: 22px
+            line-height: 22px
+            padding: 5px;
+            position: relative
+            width: 22px
